Update js for skipping questions

+++ js/index.js
@@ -14,7 +14,8 @@ if (scoreHistory) {
 } else {
     scoreHistory = [];
 };
-// console.log(scoreHistory);
+console.log(scoreHistory);
+
 let prevScore = scoreHistory[scoreHistory.length -1] || 0;
 let totalScore = prevScore;
 if (totalScoreElement) {
@@ -23,6 +24,26 @@ if (totalScoreElement) {
 let questionAnswered = false;
 let nextQuestion = document.querySelector(".activate");
 let prevQuestion = document.querySelector(".previous");
+let currentQuestion = localStorage.getItem('currentQuestion');
+console.log(currentQuestion);
+
+function checkCurrentQuestion() {
+    if (currentQuestion) {
+        currentQuestion = JSON.parse(currentQuestion);
+    } else {
+        currentQuestion = 1;
+    }
+    
+
+    if (prevQuestion) {
+        let prevQuestionNumber = currentQuestion - 1;
+        prevQuestion.href = `question${prevQuestionNumber}.html`;
+    }
+    if (nextQuestion) {
+        let nextQuestionNumber = currentQuestion + 1;
+        nextQuestion.href = `question${nextQuestionNumber}.html`;
+    }
+};
 
 // event listener to start the quiz
 if (activate) {
@@ -31,12 +52,18 @@ if (activate) {
         let agentUser = prompt('Enter Quiz takers Name');
         realUser.textContent = agentUser;
         console.log(realUser);
+        resetQuestions();
         resetScore();
     })
 }
 
 function resetScore() {
     localStorage.setItem('scoreHistory', JSON.stringify([]));
+}
+
+function resetQuestions() {
+    currentQuestion = 1;
+    localStorage.setItem('currentQuestion', JSON.stringify(currentQuestion));
 }
 
 // logic for checking answer and updating score
@@ -49,27 +76,35 @@ function updateScore(isCorrect, correctAnswer) {
         displayMsg.style.color = "green"
         score = 2;
         alert(`You got ${score} points`)
-        // console.log(score + "TQscore");
+        console.log(score + "TQscore");
         totalScore = prevScore + score
     } else {
         displayMsg.textContent = "You are wrong the correct answer is '" + correctAnswer + "'"
         displayMsg.style.color = "red"
         score = 0;
         alert(`You got ${score} points`)
-        // console.log(score + "TQscore");
+        console.log(score + "TQscore");
         totalScore = prevScore + score    
     }
     totalScoreElement.textContent = `Total Score: ${totalScore}`;
-    // console.log(totalScore + "totalScore")
+    console.log(totalScore + "totalScore")
     questionAnswered = true; 
 }
 
 // function to save score progress
-function saveProgress() {
+function saveProgress(event) {
         if (questionAnswered) {
+            if (currentQuestion === 7) {  // Replace 7 with the total number of questions currently available to answer
+                return; 
+            } 
             scoreHistory.push(totalScore);
             localStorage.setItem('scoreHistory', JSON.stringify(scoreHistory));
-            // console.log(totalScore + "saving");
+            console.log(totalScore + "saving");
+            currentQuestion++;
+            localStorage.setItem('currentQuestion', JSON.stringify(currentQuestion));
+        } else {
+            event.preventDefault();
+            alert(`Please answer Question ${currentQuestion} before continuing.`);
         }
        questionAnswered = false; 
 }
@@ -77,18 +112,29 @@ function saveProgress() {
 // score-progress
 // function to revert score progress
 function revertProgress() {
+    if (currentQuestion === 1) {   // Stop the function if the first question is reached
+        return;
+    }    
         scoreHistory.pop();
         totalScore = scoreHistory[scoreHistory.length-1] || 0;
         localStorage.setItem('scoreHistory', JSON.stringify(scoreHistory));
         questionAnswered = true;
+        currentQuestion--;
+        localStorage.setItem('currentQuestion', JSON.stringify(currentQuestion));
         window.history.back();
 }
 
 // next question button event listener to trigger saving the total score
 if (nextQuestion) {
-    nextQuestion.addEventListener("click", saveProgress);
-};
+    nextQuestion.addEventListener("click", function(event) {
+        checkCurrentQuestion();
+        saveProgress(event);
+});
+}
 // previous question button event listener to trigger reverting the total score 
 if (prevQuestion) {
-    prevQuestion.addEventListener("click", revertProgress);
-};
+    prevQuestion.addEventListener("click", function(event) {
+        checkCurrentQuestion();
+        revertProgress();
+});
+}
